refactor(svelte2tsx): スロット/イベント収集プリパスを collect/ へ切り出す
collect_template_info の walk 本体を collect/mod.rs、手書きの分割代入
パターンパーサを collect/pattern.rs へ純粋移動。出力・ロジックの変更なし。
